# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Integration Google Calendar",
        "version": "0.0.1"
    },
    "paths": {
...
        "api/google/credentials/register-key": {
            "post": {
                "tags": [
                    "Credentials Google"
                ],
                "summary": "Cria credenciais de configuração",
                "description": "A rota realiza a criação de credenciais de configurações da integração",
                "operationId": "createCredential",
                "requestBody": {
                    "description": "User credentials config google",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "google_calendar_id": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "google_client_id": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "google_client_secret": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "google_redirect_uri": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "google_webhook_uri": {
                                        "type": "string",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid credentials"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        "credential": {
                                            "$ref": "#/components/schemas/GoogleCredentials"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "api/google/events/list": {
            "get": {
                "tags": [
                    "Eventos"
                ],
                "summary": "Lista Eventos",
                "description": "A rota realiza a listagem de eventos associados ao usuário conectado",
                "operationId": "listEvents",
                "parameters": [
                    {
                        "name": "credentialId",
                        "in": "path",
                        "description": "Id da credencial",
                        "required": false
                    },
                    {
                        "name": "timeMin",
                        "in": "path",
                        "description": "Initial date filter",
                        "required": false,
                        "example": "2022-12-27 00:00:00"
                    },
                    {
                        "name": "timeMax",
                        "in": "path",
                        "description": "Finally date filter",
                        "required": false,
                        "example": "2022-12-30 00:00:00"
                    },
                    {
                        "name": "iCalUID",
                        "in": "path",
                        "description": "Id Evento",
                        "required": false
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The start date must be greater than the end date"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "[Lista com eventos]"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "google/events/show": {
            "get": {
                "tags": [
                    "Eventos"
                ],
                "summary": "Lista um evento",
                "description": "A rota realiza a exibição de um evento expecifico",
                "operationId": "showEvent",
                "parameters": [
                    {
                        "name": "credentialId",
                        "in": "path",
                        "description": "Id da credencial",
                        "required": false
                    },
                    {
                        "name": "eventId",
                        "in": "path",
                        "description": "Id Evento",
                        "required": true
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No event id provided"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "[Objecto do evento]"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "api/google/events/remove": {
            "delete": {
                "tags": [
                    "Eventos"
                ],
                "summary": "Deleta Credencial",
                "description": "A rota realiza a deleção da credencial associada ao usuário",
                "operationId": "removeEvent",
                "parameters": [
                    {
                        "name": "credentialId",
                        "in": "path",
                        "description": "Id da credencial",
                        "required": false
                    },
                    {
                        "name": "eventId",
                        "in": "path",
                        "description": "Id Evento",
                        "required": true
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No event id provided"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Event removed successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "api/google/events/create": {
            "post": {
                "tags": [
                    "Eventos"
                ],
                "summary": "Cria um evento",
                "description": "A rota realiza a criação de um evento na agenda do usuário conectado",
                "operationId": "createEvent",
                "requestBody": {
                    "description": "User credentials config google",
                    "required": true
                },
                "responses": {
                    "400": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ops, check the data and try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully created event"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "GoogleCredentials": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "google_calendar_id": {
                        "type": "text",
                        "readOnly": "true"
                    },
                    "google_client_id": {
                        "type": "text",
                        "readOnly": "true"
                    },
                    "google_client_secret": {
                        "type": "text",
                        "readOnly": "true"
                    },
                    "google_redirect_uri": {
                        "description": "Redirection route evoked by google on Oauth2 login",
                        "type": "text",
                        "readOnly": "true"
                    },
                    "google_webhook_uri": {
                        "description": "Redirection route evoked by when there is a change in the event",
                        "type": "text",
                        "readOnly": "true"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "created_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/created_at"
                    },
                    "updated_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/updated_at"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "GoogleCredentials"
                }
            },
            "User": {
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "readOnly": "true",
                        "example": "John Due"
                    },
                    "email": {
                        "description": "User unique email address",
                        "type": "string",
                        "format": "email",
                        "readOnly": "true",
                        "example": "user@gmail.com"
                    },
                    "provider_id": {
                        "type": "string",
                        "readOnly": "true",
                        "example": "1051527092716338246212"
                    },
                    "provider_name": {
                        "type": "string",
                        "readOnly": "true",
                        "example": "google"
                    },
                    "google_access_token_json": {
                        "type": "string",
                        "readOnly": "true"
                    },
                    "created_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/created_at"
                    },
                    "updated_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/updated_at"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User"
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}